feat: payggy in group chat
